Add u8 to bench

+++ benches/bench.rs
@@ -139,7 +139,8 @@ macro_rules! bench_integer_type {
 }
 
 fn benchmark(c: &mut Criterion) {
-    bench_integer_type!(c, u64);
+    // bench_integer_type!(c, u64);
+    bench_integer_type!(c, u8);
 }
 
 criterion_group!(bench, benchmark);
